requirements.txt
opencv-python==4.8.1.78
python-pptx==0.6.23
Pillow==10.1.0
numpy==1.24.3
